feat(pricing): vectorized Greek engine with CPU/CUDA backends (Levels 1-3)
Replace row-by-row iterrows() path in OptionsBase.compute_greeks() with a
fully vectorized batch_greeks() engine. CPU numpy backend is ~40-50x faster
than the scalar loop. Optional CuPy/CUDA backend for GPU acceleration at scale.

Core changes:
- core/greeks.py: add batch_greeks(), _batch_greeks_numpy(), _batch_greeks_loop(),
  _batch_greeks_cuda(), _resolve_greeks_backend(), _cupy_available(),
  _cuda_device_count(). CUDA is lazy-loaded; non-GPU machines unaffected.
- adapters/options_base.py: rewrite compute_greeks() — vectorized underlying
  precedence (underlying_price > S > F > price_std), reads greeks_backend /
  greeks_batch_size / greeks_dtype / greeks_cuda_min_rows from config.
- core/config.py: promote greeks_backend, greeks_batch_size, greeks_dtype,
  greeks_cuda_min_rows from pricing.* to root-level via normalize_config().
- run_pipeline.py: add --greeks-backend CLI flag; fix stale "row-by-row loop"
  warning in --compute-greeks help text.
- core/validators.py: fix pandas FutureWarning — cast busday_count to int64.

Tests: +31 tests (734 total, all pass, 8 skipped)
- Batch parity: Black-76 + BSM grid vs single_leg_greeks (rtol=1e-10)
- Invalid rows → NaN for all Greek outputs
- Row-order invariant + single-row perturbation locality (leakage guards)
- Backend equivalence: loop == numpy == auto
- Chunking identity, CUDA monkeypatch tests, optional GPU tests (skip if no GPU)
- Adapter: underlying precedence, future-row isolation, later-date non-leakage

Docs: docs/greek_compute_acceleration.md — full API reference, formulas,
performance table, config guide, Level 3 CUDA install instructions.
Optional deps: requirements-cuda.txt (cupy-cuda12x).

+++ core/validators.py
@@ -144,9 +144,9 @@ def missing_completeness(df: pd.DataFrame, cfg: dict) -> pd.DataFrame:
             grp = grp.sort_values("as_of_date")
             if gap_basis in {"business", "trading"}:
                 dates = pd.to_datetime(grp["as_of_date"]).dt.date.to_numpy(dtype="datetime64[D]")
-                elapsed = pd.Series(0, index=grp.index, dtype=int)
+                elapsed = pd.Series(0, index=grp.index, dtype="int64")
                 if len(dates) > 1:
-                    elapsed.iloc[1:] = np.busday_count(dates[:-1], dates[1:], holidays=holidays)
+                    elapsed.iloc[1:] = np.busday_count(dates[:-1], dates[1:], holidays=holidays).astype("int64")
                 gaps = elapsed > gap_threshold
             else:
                 gaps = grp["as_of_date"].diff().dt.days > gap_threshold
